requirements.txt

# --- Serveur Web ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# --- Base de donnees ---
sqlalchemy>=2.0.30
PyMySQL>=1.1.0
psycopg2-binary>=2.9.9

# --- Authentification ---
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# --- Formulaires et fichiers ---
python-multipart>=0.0.9

# --- Appels HTTP externes ---
httpx>=0.27.0

# --- IA Groq ---
groq>=0.9.0

# --- Utilitaires ---
python-dotenv>=1.0.1
Pillow>=10.4.0
aiofiles>=23.2.1
email-validator>=2.0.0
